refactor components for nesting server islands (fetch reviews then generate ai summary)
