+++ src/components/astro/ai-summary-island.astro
@@ -1,0 +1,33 @@
+---
+import { AISummary } from "@/components/ai-summary";
+import { summarizeReviews } from "@/lib/ai-summarizer";
+import type { Product, Review } from "@/lib/sample-data";
+
+interface Props {
+	product: Product;
+	reviews: Review[];
+}
+const { product, reviews } = Astro.props;
+
+const cacheKey = `summary-product:${product.id}` as const;
+let summary = await Astro.session?.get(cacheKey);
+
+if (!summary) {
+	try {
+		const aiResult = await summarizeReviews(product.name, reviews);
+		if (aiResult?.text) {
+			summary = {
+				text: aiResult.text,
+				timestamp: String(aiResult.response.timestamp),
+			};
+			await Astro.session?.set(cacheKey, summary);
+		} else {
+			console.warn("Failed to generate summary");
+		}
+	} catch (error) {
+		console.error("Failed to generate summary", error);
+	}
+}
+---
+
+<AISummary text={summary?.text || ""} />
