+++ src/components/astro/index.ts
@@ -1,4 +1,6 @@
+export { default as AISummaryIsland } from "./ai-summary-island.astro";
 export { default as ProductImage } from "./product-image.astro";
 export { default as ReviewsSectionIsland } from "./reviews-section-island.astro";
+export { default as ReviewsWrapper } from "./reviews-wrapper.astro";
 export { default as SiteFooter } from "./site-footer.astro";
 export { default as SiteHeader } from "./site-header.astro";
